[PULSEBS-118] refactor API.js, split into modules
