+++ client/src/api/bookingApi.js
@@ -1,0 +1,70 @@
+import Booking from './models/booking';
+import BookingExtended from './models/booking_extended';
+import axios from 'axios';
+
+const baseURL = '/api';
+
+/**
+ * Get list of bookings for a lecture
+ *
+ * @param {*} lecture_id valid lecture id
+ * @returns [] of Bookings or empty []
+ */
+async function getBookings(lecture_id){
+    let url = `/lectures/${lecture_id}/bookings`;
+
+    const response = await axios.get(baseURL + url).catch(error => {
+        if (error.response) {
+            let err = { status: error.response.status, errObj: error.response.data };
+            throw err;  // An object with the error coming from the server
+        } else if (error.request) {
+            // The request was made but no response was received
+            console.log(error.request);
+        } else {
+            // Something happened in setting up the request that triggered an Error
+            console.log('Error', error.message);
+        }
+    });
+    if (response.status === 200) {
+        return response.data.map(
+            (o) => new BookingExtended(o.lecture_id, o.student_id, o.waiting, o.present, o.updated_at, o.deleted_at, o.student_university_id, o.student_name, o.student_surname));
+    } else {
+        let err = { status: response.status, errObj: response.data };
+        throw err;  // An object with the error coming from the server
+    }
+}
+
+/**
+ * Book a seat to a lecture given a student id and lecture id
+ *
+ * @param {*} student_id a valid student id
+ * @param {*} lecture_id a valid lecture id
+ * @returns the newly created Booking object
+ */
+async function bookLecture(student_id, lecture_id){
+    let url = `/students/${student_id}/bookings`;
+
+    const response = await axios.post(baseURL + url, {
+        lecture_id: lecture_id
+    }).catch(error => {
+        if (error.response) {
+            let err = { status: error.response.status, errObj: error.response.data };
+            throw err;  // An object with the error coming from the server
+        } else if (error.request) {
+            // The request was made but no response was received
+            console.log(error.request);
+        } else {
+            // Something happened in setting up the request that triggered an Error
+            console.log('Error', error.message);
+        }
+    });
+    if (response.status === 201) {
+        const booking = response.data;
+        return new Booking(booking.lecture_id, booking.student_id, booking.waiting, booking.present, booking.updated_at, booking.deleted_at);
+    } else {
+        let err = { status: response.status, errObj: response.data };
+        throw err;  // An object with the error coming from the server
+    }
+}
+
+export {bookLecture, getBookings}
